Refactor: General tab functionality to dockwidget module.
